Refactoring and moving decorator logic to __init__
