+++ ckanext/bcgov/logic/ofi/call_action.py
@@ -1,5 +1,11 @@
+# Copyright 2016, Province of British Columbia
+# License: https://github.com/bcgov/ckanext-bcgov/blob/master/license
+#
+# Author: Jared Smith <jrods@github>
+# HighwayThree Solutions Inc.
+
 import logging
-from functools import wraps
+
 from pprint import pprint, pformat
 
 import requests as reqs
@@ -7,31 +13,15 @@ import requests as reqs
 import ckan.logic as logic
 from ckan.common import _, c, request, response
 
+import ckanext.bcgov.logic.ofi as ofi_logic
 import ckanext.bcgov.util.helpers as edc_h
 
 
 log = logging.getLogger(u'ckanext.bcgov.logic.ofi')
 
 
-def setup_action(api_url):
-    '''
-    Decorator for call_action functions, macro for setting up all the vars for ofi call
-    '''
-    def action_decorator(action):
-        @wraps(action)
-        def wrapper(context, data):
-            '''
-            Context and data are args for get_action calls
-            '''
-            ofi_vars = _prepare(data[u'secure'])
-            url = ofi_vars[u'ofi_url'] + api_url + data[u'object_name']
-            call_type = 'Secure' if data[u'secure'] else 'Public'  # call_type is for logging purposes
-            return action(url, ofi_vars, call_type)
-        return wrapper
-    return action_decorator
-
-
-@setup_action(u'/info/fileFormats')
+@logic.side_effect_free
+@ofi_logic.setup_action(u'/info/fileFormats')
 def file_formats(url, ofi_vars, call_type):
     '''
     OFI API Call:
@@ -41,7 +31,7 @@ def file_formats(url, ofi_vars, call_type):
 
 
 @logic.side_effect_free
-@setup_action(u'/security/productAllowedByFeatureType/')
+@ofi_logic.setup_action(u'/security/productAllowedByFeatureType/')
 def check_object_name(url, ofi_vars, call_type):
     '''
     OFI API Call:
@@ -64,24 +54,6 @@ def _make_api_call(api_url, call_type='Public', cookies=None):
         return u'Need valid IDIR Login session'
     else:
         return resp.json()
-
-
-def _prepare(secure=False):
-    ofi_vars = {}
-    ofi_vars[u'config'] = edc_h.get_ofi_config()
-    ofi_vars[u'cookies'] = {
-        u'SMSESSION': request.cookies.get(u'SMSESSION', '')
-    }
-    try:
-        ofi_vars['query_params'] = request.params
-    except ValueError, inst:
-        log.info('Bad Action API request data: %s', inst)
-        return {}
-
-    ofi_vars[u'secure'] = secure
-    ofi_vars[u'ofi_url'] = edc_h._build_url(secure)
-
-    return ofi_vars
 
 
 def _log_response(resp, call_type):
